Remove duplicate translations
Added a command that helps detecting duplicates, the file became too big for Cursor to process so everytime there is a duplicate error it creates a new script so it is a waste of tokens. There is a command that detects them. Just ask Cursor to execute it everytime there is a new translation and that is it.